Linux changes

+++ install/arch.sh
@@ -2,23 +2,29 @@
 
 sudo pacman --noconfirm -Suyy git base-devel
 
-git clone https://aur.archlinux.org/yay-bin.git /tmp/yay-bin
-cd /tmp/yay-bin
-makepkg -is
-cd -
-rm -rf /tmp/yay-bin
+if ! [ -x "$(command -v yay)" ]; then
+	git clone https://aur.archlinux.org/yay-bin.git /tmp/yay-bin
+	cd /tmp/yay-bin
+	makepkg -is
+	cd -
+	rm -rf /tmp/yay-bin
+fi
 
 # common
 yay --needed --noconfirm -S neovim tmux zsh unzip zip efibootmgr wmname dunst \
 	network-manager-applet htop filelight playerctl \
-	telegram-desktop spotify spotifywm-git bitwarden-bin bc \
+	lazygit git-crypt bolt eza \
+	telegram-desktop spotifywm-git bitwarden-bin bc \
 	jetbrains-toolbox nerd-fonts-iosevka \
 	ttf-iosevka-aile insync bluez pavucontrol paprefs \
 	android-tools android-udev android-sdk spicetify-cli spicetify-themes-git \
-	brightnessctl firefox blueman keyd-git exa jq go fzf xxd wezterm
+	brightnessctl firefox blueman keyd-git jq go fzf xxd wezterm
+
+# this shit has GPG keys often failing to import
+yay --needed --noconfirm -S spotify
 
 # pipewire version
-yay --needed --noconfirm -S pipewire pipewire-pulse pipewire-jack-dropin pipewire-jack pipewire-media-session
+yay --needed --noconfirm -S pipewire pipewire-pulse pipewire-jack-dropin pipewire-jack wireplumber
 
 # Hyprland version
 yay --needed --noconfirm -S rofi-wayland waybar slurp grim swappy wtype wl-clipboard hyprland hy3
